[BTAT-2248] Refactored the LastTaxCalculation model into latestTaxCalculation package. Update IT tests
