+++ it/assets/BusinessDetailsIntegrationTestConstants.scala
@@ -21,7 +21,6 @@ import java.time.LocalDate
 import assets.BaseIntegrationTestConstants.testMtdRef
 import models.core.{AccountingPeriodModel, AddressModel, ContactDetailsModel}
 import models.incomeSourceDetails.BusinessDetailsModel
-import models.{AddressModel, ContactDetailsModel}
 import play.api.libs.json.{JsValue, Json}
 
 object BusinessDetailsIntegrationTestConstants {
