+++ test/mocks/MockEstimatedTaxLiabilityService.scala
@@ -18,7 +18,7 @@ package mocks
 
 import assets.FinancialDataTestConstants.{testCalcType, testYear}
 import assets.BaseTestConstants.testNino
-import models.LastTaxCalculationResponseModel
+import models.latestTaxCalculation.LastTaxCalculationResponseModel
 import org.mockito.ArgumentMatchers
 import org.mockito.Mockito._
 import org.mockito.stubbing.OngoingStubbing
