+++ test/mocks/MockFinancialDataConnector.scala
@@ -16,10 +16,10 @@
 
 package mocks
 
-import assets.FinancialDataTestConstants.{testYear,testCalcType}
+import assets.FinancialDataTestConstants.{testCalcType, testYear}
 import assets.BaseTestConstants.testNino
 import connectors.FinancialDataConnector
-import models.LastTaxCalculationResponseModel
+import models.latestTaxCalculation.LastTaxCalculationResponseModel
 import org.mockito.ArgumentMatchers
 import org.mockito.Mockito._
 import org.mockito.stubbing.OngoingStubbing
